refactor common code into common.py to prevent circular imports when passing decode_times and decode_timedelta to CFMaskCoder and CFScaleOffsetCoder
