+++ xarray/coding/variables.py
@@ -3,85 +3,29 @@
 from __future__ import annotations
 
 import warnings
-from collections.abc import Callable, Hashable, MutableMapping
+from collections.abc import Hashable, MutableMapping
 from functools import partial
 from typing import TYPE_CHECKING, Any, Union
 
 import numpy as np
 import pandas as pd
 
+from xarray.coding.common import (
+    SerializationWarning,
+    VariableCoder,
+    lazy_elemwise_func,
+    pop_to,
+    safe_setitem,
+    unpack_for_decoding,
+    unpack_for_encoding,
+)
+from xarray.coding.times import CFDatetimeCoder, CFTimedeltaCoder
 from xarray.core import dtypes, duck_array_ops, indexing
 from xarray.core.variable import Variable
-from xarray.namedarray.parallelcompat import get_chunked_array_type
-from xarray.namedarray.pycompat import is_chunked_array
 
 if TYPE_CHECKING:
     T_VarTuple = tuple[tuple[Hashable, ...], Any, dict, dict]
     T_Name = Union[Hashable, None]
-
-
-class SerializationWarning(RuntimeWarning):
-    """Warnings about encoding/decoding issues in serialization."""
-
-
-class VariableCoder:
-    """Base class for encoding and decoding transformations on variables.
-
-    We use coders for transforming variables between xarray's data model and
-    a format suitable for serialization. For example, coders apply CF
-    conventions for how data should be represented in netCDF files.
-
-    Subclasses should implement encode() and decode(), which should satisfy
-    the identity ``coder.decode(coder.encode(variable)) == variable``. If any
-    options are necessary, they should be implemented as arguments to the
-    __init__ method.
-
-    The optional name argument to encode() and decode() exists solely for the
-    sake of better error messages, and should correspond to the name of
-    variables in the underlying store.
-    """
-
-    def encode(self, variable: Variable, name: T_Name = None) -> Variable:
-        """Convert an encoded variable to a decoded variable"""
-        raise NotImplementedError()
-
-    def decode(self, variable: Variable, name: T_Name = None) -> Variable:
-        """Convert a decoded variable to an encoded variable"""
-        raise NotImplementedError()
-
-
-class _ElementwiseFunctionArray(indexing.ExplicitlyIndexedNDArrayMixin):
-    """Lazily computed array holding values of elemwise-function.
-
-    Do not construct this object directly: call lazy_elemwise_func instead.
-
-    Values are computed upon indexing or coercion to a NumPy array.
-    """
-
-    def __init__(self, array, func: Callable, dtype: np.typing.DTypeLike):
-        assert not is_chunked_array(array)
-        self.array = indexing.as_indexable(array)
-        self.func = func
-        self._dtype = dtype
-
-    @property
-    def dtype(self) -> np.dtype:
-        return np.dtype(self._dtype)
-
-    def _oindex_get(self, key):
-        return type(self)(self.array.oindex[key], self.func, self.dtype)
-
-    def _vindex_get(self, key):
-        return type(self)(self.array.vindex[key], self.func, self.dtype)
-
-    def __getitem__(self, key):
-        return type(self)(self.array[key], self.func, self.dtype)
-
-    def get_duck_array(self):
-        return self.func(self.array.get_duck_array())
-
-    def __repr__(self) -> str:
-        return f"{type(self).__name__}({self.array!r}, func={self.func!r}, dtype={self.dtype!r})"
 
 
 class NativeEndiannessArray(indexing.ExplicitlyIndexedNDArrayMixin):
@@ -159,63 +103,6 @@ class BoolTypeArray(indexing.ExplicitlyIndexedNDArrayMixin):
 
     def __getitem__(self, key) -> np.ndarray:
         return np.asarray(self.array[key], dtype=self.dtype)
-
-
-def lazy_elemwise_func(array, func: Callable, dtype: np.typing.DTypeLike):
-    """Lazily apply an element-wise function to an array.
-    Parameters
-    ----------
-    array : any valid value of Variable._data
-    func : callable
-        Function to apply to indexed slices of an array. For use with dask,
-        this should be a pickle-able object.
-    dtype : coercible to np.dtype
-        Dtype for the result of this function.
-
-    Returns
-    -------
-    Either a dask.array.Array or _ElementwiseFunctionArray.
-    """
-    if is_chunked_array(array):
-        chunkmanager = get_chunked_array_type(array)
-
-        return chunkmanager.map_blocks(func, array, dtype=dtype)  # type: ignore[arg-type]
-    else:
-        return _ElementwiseFunctionArray(array, func, dtype)
-
-
-def unpack_for_encoding(var: Variable) -> T_VarTuple:
-    return var.dims, var.data, var.attrs.copy(), var.encoding.copy()
-
-
-def unpack_for_decoding(var: Variable) -> T_VarTuple:
-    return var.dims, var._data, var.attrs.copy(), var.encoding.copy()
-
-
-def safe_setitem(dest, key: Hashable, value, name: T_Name = None):
-    if key in dest:
-        var_str = f" on variable {name!r}" if name else ""
-        raise ValueError(
-            f"failed to prevent overwriting existing key {key} in attrs{var_str}. "
-            "This is probably an encoding field used by xarray to describe "
-            "how a variable is serialized. To proceed, remove this key from "
-            "the variable's attributes manually."
-        )
-    dest[key] = value
-
-
-def pop_to(
-    source: MutableMapping, dest: MutableMapping, key: Hashable, name: T_Name = None
-) -> Any:
-    """
-    A convenience function which pops a key k from source to dest.
-    None values are not passed on.  If k already exists in dest an
-    error is raised.
-    """
-    value = source.pop(key, None)
-    if value is not None:
-        safe_setitem(dest, key, value, name=name)
-    return value
 
 
 def _apply_mask(
@@ -371,8 +258,8 @@ class CFMaskCoder(VariableCoder):
 
     def __init__(
         self,
-        decode_times: bool = False,
-        decode_timedelta: bool = False,
+        decode_times: bool | CFDatetimeCoder = False,
+        decode_timedelta: bool | CFTimedeltaCoder = False,
     ) -> None:
         self.decode_times = decode_times
         self.decode_timedelta = decode_timedelta
@@ -407,9 +294,11 @@ class CFMaskCoder(VariableCoder):
             encoding["_FillValue"] = (
                 _encode_unsigned_fill_value(name, fv, dtype)
                 if has_unsigned
-                else dtype.type(fv)
-                if "add_offset" not in encoding and "scale_factor" not in encoding
-                else fv
+                else (
+                    dtype.type(fv)
+                    if "add_offset" not in encoding and "scale_factor" not in encoding
+                    else fv
+                )
             )
             fill_value = pop_to(encoding, attrs, "_FillValue", name=name)
 
@@ -422,9 +311,12 @@ class CFMaskCoder(VariableCoder):
                 (
                     _encode_unsigned_fill_value(name, mv, dtype)
                     if has_unsigned
-                    else dtype.type(mv)
-                    if "add_offset" not in encoding and "scale_factor" not in encoding
-                    else mv
+                    else (
+                        dtype.type(mv)
+                        if "add_offset" not in encoding
+                        and "scale_factor" not in encoding
+                        else mv
+                    )
                 ),
             )
             fill_value = pop_to(encoding, attrs, "missing_value", name=name)
@@ -587,8 +479,8 @@ class CFScaleOffsetCoder(VariableCoder):
 
     def __init__(
         self,
-        decode_times: bool = False,
-        decode_timedelta: bool = False,
+        decode_times: bool | CFDatetimeCoder = False,
+        decode_timedelta: bool | CFTimedeltaCoder = False,
     ) -> None:
         self.decode_times = decode_times
         self.decode_timedelta = decode_timedelta
